+++ src/main/java/kono/ceu/materialreplication/api/unification/materials/MRMaterial.java
@@ -13,16 +13,22 @@ public class MRMaterial {
                 .color(0xc86edc)
                 .iconSet(MaterialIconSet.NETHERSTAR).build();
 
-        MRMaterials.ChargedMatter = new Material.Builder(32008, "charged_matter")
+        MRMaterials.ChargedMatter = new Material.Builder(31001, "charged_matter")
                 .fluid()
                 .fluidTemp(1)
                 .color(0x0064c8)
                 .build();
 
-        MRMaterials.NeutralMatter = new Material.Builder(32009, "neutral_matter")
+        MRMaterials.NeutralMatter = new Material.Builder(31002, "neutral_matter")
                 .fluid()
                 .fluidTemp(1)
                 .color(0xc80a14)
+                .build();
+
+        MRMaterials.MatterAmplifier = new Material.Builder(31003, "matter_amplifier")
+                .dust().fluid()
+                .fluidTemp(300)
+                .color(0xda70d6)
                 .build();
     }
 }