+++ src/main/java/kono/ceu/materialreplication/common/items/MRMetaItem1.java
@@ -1,6 +1,7 @@
 package kono.ceu.materialreplication.common.items;
 
 import gregtech.api.items.metaitem.StandardMetaItem;
+import kono.ceu.materialreplication.common.items.behaviors.ReplicationBehaviors;
 
 import static kono.ceu.materialreplication.common.items.MRMetaItems.*;
 
@@ -12,6 +13,8 @@ public class MRMetaItem1 extends StandardMetaItem {
 
     @Override
     public void registerSubItems() {
-        USB_STICK = addItem(0, "usb_stick");
+        USB_STICK = addItem(0, "usb_stick").addComponents(new ReplicationBehaviors());
+        SCRAP = addItem(1, "scrap");
+        SCRAP_BOX = addItem(2, "scrap_box");
     }
 }