+++ src/main/java/kono/ceu/materialreplication/loaders/recipe/MRMachineRecipeLoader.java
@@ -2,39 +2,40 @@ package kono.ceu.materialreplication.loaders.recipe;
 
 import gregtech.api.GregTechAPI;
 import gregtech.api.recipes.RecipeBuilder;
-import gregtech.api.recipes.builders.SimpleRecipeBuilder;
+import gregtech.api.recipes.ingredients.GTRecipeItemInput;
+import gregtech.api.recipes.ingredients.nbtmatch.NBTCondition;
+import gregtech.api.recipes.ingredients.nbtmatch.NBTMatcher;
+import gregtech.api.recipes.ingredients.nbtmatch.NBTTagType;
 import gregtech.api.unification.material.Material;
 import gregtech.api.unification.material.properties.PropertyKey;
-import kono.ceu.materialreplication.MRConfig;
 import kono.ceu.materialreplication.api.recipes.MRRecipeMaps;
+import kono.ceu.materialreplication.api.recipes.builders.ReplicatorRecipeBuilder;
+import kono.ceu.materialreplication.api.recipes.machines.IReplicatorRecipeMap;
 import kono.ceu.materialreplication.api.unification.materials.MRMaterials;
 import kono.ceu.materialreplication.api.unification.materials.flags.MRMaterialFlags;
+import kono.ceu.materialreplication.common.items.MRMetaItems;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import static gregtech.api.unification.ore.OrePrefix.*;
+import static gregtech.api.GTValues.*;
+import static gregtech.api.recipes.RecipeMaps.*;
+import static gregtech.api.unification.ore.OrePrefix.dust;
+import static gregtech.api.unification.ore.OrePrefix.dustTiny;
+import static kono.ceu.materialreplication.api.MRValues.*;
+import static kono.ceu.materialreplication.common.items.MRMetaItems.SCRAP;
+import static kono.ceu.materialreplication.common.items.MRMetaItems.SCRAP_BOX;
 
 public class MRMachineRecipeLoader {
+    public static void register() {
+        recipeMatter();
+        recipeScrap();
+    }
 
-    // Get Deconstruction Base-time from cfg. cfgから分解の基本時間を取得
-    private final static int BaseTime_D = MRConfig.deconstruction.DeconstructionBaseTime;
-
-    // Get Deconstruction Base Voltage from cfg. cfgから分解の基本電圧を取得
-    private final static int Voltage_D = MRConfig.deconstruction.DeconstructionVoltage;
-
-    // Get Replication Base-time from cfg. cfgから複製の基本時間を取得
-    private final static int BaseTime_R = MRConfig.replication.ReplicationBaseTime;
-
-    // Get Replication Base Voltage from cfg. cfgから複製の基本電圧を取得
-    private final static int Voltage_R = MRConfig.replication.ReplicationVoltage;
-
-
-
-    public static void init() {
+    public static void recipeMatter() {
         List<Material> materialDusts = new ArrayList<>();
         List<Material> materialFluids = new ArrayList<>();
-        for (Material material : GregTechAPI.MATERIAL_REGISTRY) {
+       for (Material material : GregTechAPI.MATERIAL_REGISTRY) {
             if (!material.getChemicalFormula().isEmpty()) { // Has Chemical Formula? 化学式を持っているか
                 if (material.hasProperty(PropertyKey.DUST)) { // Has Dust Property? Propertyにdustがあるか
                     materialDusts.add(material);
@@ -44,7 +45,8 @@ public class MRMachineRecipeLoader {
             }
         }
 
-        for (Material materialDust : materialDusts) { // Has dust property material. dustがあるmaterial
+        for (Material materialDust : materialDusts) {
+            // Has dust property material. dustがあるmaterial
             // Deconstruction
             if (!materialDust.hasFlag(MRMaterialFlags.DISABLE_DECONSTRUCTION)) {
                 MRRecipeMaps.DECONSTRUCTION_RECIPES.recipeBuilder()
@@ -59,52 +61,30 @@ public class MRMachineRecipeLoader {
 
             // Replication
             if (!materialDust.hasFlag(MRMaterialFlags.DISABLE_REPLICATION)) {
+                RecipeBuilder<ReplicatorRecipeBuilder> builder = MRRecipeMaps.REPLICATION_RECIPES.recipeBuilder();
                 if (materialDust.getProtons() == 0) {
-                    RecipeBuilder<SimpleRecipeBuilder> builder = MRRecipeMaps.REPLICATION_RECIPES.recipeBuilder()
-                            .fluidInputs(MRMaterials.NeutralMatter.getFluid((int) materialDust.getNeutrons()))
-                            .duration(BaseTime_R * (int) materialDust.getMass())
-                            .EUt(Voltage_R);
-                    if (MRConfig.replication.ConsumeSample) {
-                        builder.input(dust, materialDust)
-                                .output(dust, materialDust, 2);
-                    } else {
-                        builder.notConsumable(dust, materialDust)
-                                .output(dust, materialDust, 1);
-                    }
-                    builder.buildAndRegister();
+                    builder.fluidInputs(MRMaterials.NeutralMatter.getFluid((int) materialDust.getNeutrons()));
                 } else if (materialDust.getNeutrons() == 0) {
-                    RecipeBuilder<SimpleRecipeBuilder> builder = MRRecipeMaps.REPLICATION_RECIPES.recipeBuilder()
-                            .fluidInputs(MRMaterials.ChargedMatter.getFluid((int) materialDust.getProtons()))
-                            .duration(BaseTime_R * (int) materialDust.getMass())
-                            .EUt(Voltage_R);
-                    if (MRConfig.replication.ConsumeSample) {
-                        builder.input(dust, materialDust)
-                                .output(dust, materialDust, 2);
-                    } else {
-                        builder.notConsumable(dust, materialDust)
-                                .output(dust, materialDust, 1);
-                    }
-                    builder.buildAndRegister();
+                    builder.fluidInputs(MRMaterials.ChargedMatter.getFluid((int) materialDust.getProtons()));
                 } else {
-                    RecipeBuilder<SimpleRecipeBuilder> builder = MRRecipeMaps.REPLICATION_RECIPES.recipeBuilder()
-                            .fluidInputs(MRMaterials.ChargedMatter.getFluid((int) materialDust.getProtons()))
-                            .fluidInputs(MRMaterials.NeutralMatter.getFluid((int) materialDust.getNeutrons()))
-                            .duration(BaseTime_R * (int) materialDust.getMass())
-                            .EUt(Voltage_R);
-                    if (MRConfig.replication.ConsumeSample) {
-                        builder.input(dust, materialDust)
-                                .output(dust, materialDust, 2);
-                    } else {
-                        builder.notConsumable(dust, materialDust)
-                                .output(dust, materialDust, 1);
+                    builder.fluidInputs(MRMaterials.ChargedMatter.getFluid((int) materialDust.getProtons()))
+                            .fluidInputs(MRMaterials.NeutralMatter.getFluid((int) materialDust.getNeutrons()));
                     }
-                    builder.buildAndRegister();
+                builder.duration(BaseTime_R * (int) materialDust.getMass())
+                        .input(GTRecipeItemInput.getOrCreate(MRMetaItems.USB_STICK.getStackForm())
+                                .setNBTMatchingCondition(NBTMatcher.RECURSIVE_EQUAL_TO,
+                                        NBTCondition.create(NBTTagType.COMPOUND, IReplicatorRecipeMap.REPLICATE_NBT_TAG,
+                                                NBTCondition.create(NBTTagType.STRING, IReplicatorRecipeMap.REPLICATE_MATERIAL, materialDust.toString())))
+                                .setNonConsumable())
+                        .replicate(materialDust)
+                        .EUt(Voltage_R)
+                        .output(dust, materialDust)
+                        .buildAndRegister();
                 }
             }
 
-        }
-
-        for (Material materialFluid : materialFluids) { //Has Fluid property, not dust. dustはないがFluidはある
+        for (Material materialFluid : materialFluids) {
+            //Has Fluid property, not dust. dustはないがFluidはある
             // Deconstruction
             if (!materialFluid.hasFlag(MRMaterialFlags.DISABLE_DECONSTRUCTION)) {
                 MRRecipeMaps.DECONSTRUCTION_RECIPES.recipeBuilder()
@@ -118,49 +98,74 @@ public class MRMachineRecipeLoader {
             }
             // Replication
             if (!materialFluid.hasFlag(MRMaterialFlags.DISABLE_REPLICATION)) {
+                RecipeBuilder <ReplicatorRecipeBuilder> builder = MRRecipeMaps.REPLICATION_RECIPES.recipeBuilder();
                 if (materialFluid.getProtons() == 0) {
-                    RecipeBuilder <SimpleRecipeBuilder> builder = MRRecipeMaps.REPLICATION_RECIPES.recipeBuilder()
-                            .fluidInputs(MRMaterials.NeutralMatter.getFluid((int) materialFluid.getNeutrons()))
-                            .duration(BaseTime_R * (int) materialFluid.getMass())
-                            .EUt(Voltage_R);
-                    if (MRConfig.replication.ConsumeSample) {
-                        builder.fluidInputs(materialFluid.getFluid(1000))
-                                .fluidOutputs(materialFluid.getFluid(2000));
-                    } else {
-                        builder.notConsumable(materialFluid.getFluid(), 1000)
-                                .fluidOutputs(materialFluid.getFluid(1000));
-                    }
-                    builder.buildAndRegister();
+                    builder.fluidInputs(MRMaterials.NeutralMatter.getFluid((int) materialFluid.getNeutrons()));
                 } else if (materialFluid.getNeutrons() == 0) {
-                    RecipeBuilder <SimpleRecipeBuilder> builder = MRRecipeMaps.REPLICATION_RECIPES.recipeBuilder()
-                            .fluidInputs(MRMaterials.ChargedMatter.getFluid((int) materialFluid.getProtons()))
-                            .duration(BaseTime_R * (int) materialFluid.getMass())
-                            .EUt(Voltage_R);
-                    if (MRConfig.replication.ConsumeSample) {
-                        builder.fluidInputs(materialFluid.getFluid(1000))
-                                .fluidOutputs(materialFluid.getFluid(2000));
-                    } else {
-                        builder.notConsumable(materialFluid.getFluid(), 1000)
-                                .fluidOutputs(materialFluid.getFluid(1000));
-                    }
-                    builder.buildAndRegister();
+                    builder.fluidInputs(MRMaterials.ChargedMatter.getFluid((int) materialFluid.getProtons()));
                 } else {
-                    RecipeBuilder <SimpleRecipeBuilder> builder = MRRecipeMaps.REPLICATION_RECIPES.recipeBuilder()
-                            .fluidInputs(MRMaterials.ChargedMatter.getFluid((int) materialFluid.getProtons()))
-                            .fluidInputs(MRMaterials.NeutralMatter.getFluid((int) materialFluid.getNeutrons()))
-                            .duration(BaseTime_R * (int) materialFluid.getMass())
-                            .EUt(Voltage_R);
-                    if (MRConfig.replication.ConsumeSample) {
-                        builder.fluidInputs(materialFluid.getFluid(1000))
-                                .fluidOutputs(materialFluid.getFluid(2000));
-                    } else {
-                        builder.notConsumable(materialFluid.getFluid(), 1000)
-                                .fluidOutputs(materialFluid.getFluid(1000));
-                    }
-                    builder.buildAndRegister();
+                    builder.fluidInputs(MRMaterials.ChargedMatter.getFluid((int) materialFluid.getProtons()))
+                            .fluidInputs(MRMaterials.NeutralMatter.getFluid((int) materialFluid.getNeutrons()));
                 }
+                builder.input(GTRecipeItemInput.getOrCreate(MRMetaItems.USB_STICK.getStackForm())
+                        .setNBTMatchingCondition(NBTMatcher.RECURSIVE_EQUAL_TO,
+                                NBTCondition.create(NBTTagType.COMPOUND, IReplicatorRecipeMap.REPLICATE_NBT_TAG,
+                                        NBTCondition.create(NBTTagType.STRING, IReplicatorRecipeMap.REPLICATE_MATERIAL, materialFluid.toString())))
+                        .setNonConsumable())
+                        .fluidOutputs(materialFluid.getFluid(1000))
+                        .replicate(materialFluid)
+                        .duration(BaseTime_R * (int) materialFluid.getMass())
+                        .EUt(Voltage_R)
+                        .buildAndRegister();
             }
         }
 
+        // Primal -> Charged & Neutral
+        CENTRIFUGE_RECIPES.recipeBuilder()
+                .fluidInputs(MRMaterials.PrimalMatter.getFluid(1))
+                .fluidOutputs(MRMaterials.ChargedMatter.getFluid(1))
+                .fluidOutputs(MRMaterials.NeutralMatter.getFluid(1))
+                .duration(12000).EUt(16).buildAndRegister();
+
+        // MatterAmplifier
+        CHEMICAL_RECIPES.recipeBuilder()
+                .fluidInputs(MRMaterials.MatterAmplifier.getFluid(500))
+                .fluidInputs(MRMaterials.ChargedMatter.getFluid(500))
+                .fluidOutputs(MRMaterials.ChargedMatter.getFluid(1000))
+                .duration(BaseTime_R).EUt(Voltage_R).buildAndRegister();
+
+        CHEMICAL_RECIPES.recipeBuilder()
+                .fluidInputs(MRMaterials.MatterAmplifier.getFluid(500))
+                .fluidInputs(MRMaterials.NeutralMatter.getFluid(500))
+                .fluidOutputs(MRMaterials.NeutralMatter.getFluid(1000))
+                .duration(BaseTime_R).EUt(Voltage_R).buildAndRegister();
     }
+
+    public static void recipeScrap() {
+        COMPRESSOR_RECIPES.recipeBuilder()
+                .input(SCRAP, 9)
+                .output(SCRAP_BOX)
+                .duration(1200).EUt(VA[HV]).buildAndRegister();
+
+        SIFTER_RECIPES.recipeBuilder()
+                .input(SCRAP)
+                .chancedOutput(dustTiny, MRMaterials.MatterAmplifier, AmplifierChance, AmplifierChanceBoost)
+                .duration(256).EUt(VA[LV]).buildAndRegister();
+
+        SIFTER_RECIPES.recipeBuilder()
+                .input(SCRAP_BOX)
+                .chancedOutput(dustTiny, MRMaterials.MatterAmplifier, 2, AmplifierChance, AmplifierChanceBoost)
+                .chancedOutput(dustTiny, MRMaterials.MatterAmplifier, 2, AmplifierChance, AmplifierChanceBoost)
+                .chancedOutput(dustTiny, MRMaterials.MatterAmplifier, 2, AmplifierChance, AmplifierChanceBoost)
+                .chancedOutput(dustTiny, MRMaterials.MatterAmplifier, 1, AmplifierChance, AmplifierChanceBoost)
+                .chancedOutput(dustTiny, MRMaterials.MatterAmplifier, 1, AmplifierChance, AmplifierChanceBoost)
+                .chancedOutput(dustTiny, MRMaterials.MatterAmplifier, 1, AmplifierChance, AmplifierChanceBoost)
+                .duration(256).EUt(VA[LV]).buildAndRegister();
+
+        EXTRACTOR_RECIPES.recipeBuilder()
+                .input(dust, MRMaterials.MatterAmplifier)
+                .fluidOutputs(MRMaterials.MatterAmplifier.getFluid(10))
+                .duration(BaseTime_D / 4).EUt(VA[EV]).buildAndRegister();
+    }
+
 }